+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased](https://git.sr.ht/~rjarry/aerc/log/master)
 
+### Added
+
+- New `flagged` criteria for `:sort`
+
 ### Fixed
 
 - `colorize` styles can include wildcards `?` and `*`